Move post button to aside
